warning fixes for operators directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23663 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/operators/ConnectedComponents/avtConnectedComponentsFilter.C
@@ -161,7 +161,7 @@ avtConnectedComponentsFilter::InferVariableNameFromContract(avtContract_p c)
     if (!foundIt)
     {
         std::vector<CharStrRef>   vars2nd = in_dr->GetSecondaryVariablesWithoutDuplicates();
-        for (int i = 0 ; i < vars2nd.size() ; i++)
+        for (size_t i = 0 ; i < vars2nd.size() ; i++)
         if (strncmp(*(vars2nd[i]), "operators/ConnectedComponents/",
                     strlen("operators/ConnectedComponents/")) == 0)
         {
